move stuff out of org.hibernate.query.procedure
to realize consistent SPI/API split
