+++ hibernate-core/src/main/java/org/hibernate/procedure/ProcedureOutputs.java
@@ -4,7 +4,6 @@
  */
 package org.hibernate.procedure;
 
-import org.hibernate.query.procedure.ProcedureParameter;
 import org.hibernate.result.Outputs;
 
 /**
