+++ hibernate-core/src/main/java/org/hibernate/procedure/internal/ProcedureOutputsImpl.java
@@ -12,7 +12,7 @@ import java.util.Map;
 
 import org.hibernate.procedure.ParameterMisuseException;
 import org.hibernate.procedure.ProcedureOutputs;
-import org.hibernate.query.procedure.ProcedureParameter;
+import org.hibernate.procedure.ProcedureParameter;
 import org.hibernate.result.Output;
 import org.hibernate.result.internal.OutputsImpl;
 import org.hibernate.sql.exec.ExecutionException;
