+++ hibernate-core/src/main/java/org/hibernate/procedure/internal/ProcedureParameterBindingImpl.java
@@ -5,10 +5,9 @@
 package org.hibernate.procedure.internal;
 
 import org.hibernate.engine.spi.SessionFactoryImplementor;
-import org.hibernate.procedure.spi.ProcedureParameterBindingImplementor;
+import org.hibernate.procedure.spi.ProcedureParameterBinding;
 import org.hibernate.procedure.spi.ProcedureParameterImplementor;
 import org.hibernate.query.internal.QueryParameterBindingImpl;
-import org.hibernate.query.procedure.ProcedureParameterBinding;
 
 /**
  * Implementation of the {@link ProcedureParameterBinding} contract.
@@ -17,7 +16,7 @@ import org.hibernate.query.procedure.ProcedureParameterBinding;
  */
 public class ProcedureParameterBindingImpl<T>
 		extends QueryParameterBindingImpl<T>
-		implements ProcedureParameterBindingImplementor<T> {
+		implements ProcedureParameterBinding<T> {
 	public ProcedureParameterBindingImpl(
 			ProcedureParameterImplementor<T> queryParameter,
 			SessionFactoryImplementor sessionFactory) {
