--- hibernate-core/src/main/java/org/hibernate/query/procedure/package-info.java
@@ -1,14 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- * Copyright Red Hat Inc. and Hibernate Authors
- */
-
-/**
- * Support SQL functions in the SQM model.
- *
- * @see org.hibernate.query.sqm.function.SqmFunctionDescriptor
- */
-@Incubating
-package org.hibernate.query.procedure;
-
-import org.hibernate.Incubating;
