+++ hibernate-core/src/test/java/org/hibernate/orm/test/procedure/OracleStoredProcedureTest.java
@@ -20,7 +20,7 @@ import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.dialect.OracleDialect;
 import org.hibernate.jpa.HibernateHints;
 import org.hibernate.procedure.ProcedureCall;
-import org.hibernate.query.procedure.ProcedureParameter;
+import org.hibernate.procedure.ProcedureParameter;
 import org.hibernate.result.Output;
 import org.hibernate.result.ResultSetOutput;
 import org.hibernate.type.NumericBooleanConverter;
