+++ hibernate-core/src/test/java/org/hibernate/orm/test/sql/MySQLStoredProcedureTest.java
@@ -17,7 +17,7 @@ import org.hibernate.Session;
 import org.hibernate.dialect.MySQLDialect;
 import org.hibernate.community.dialect.TiDBDialect;
 import org.hibernate.procedure.ProcedureCall;
-import org.hibernate.query.procedure.ProcedureParameter;
+import org.hibernate.procedure.ProcedureParameter;
 import org.hibernate.result.Output;
 import org.hibernate.result.ResultSetOutput;
 import org.hibernate.testing.orm.domain.userguide.Account;
