+++ hibernate-core/src/test/java/org/hibernate/orm/test/sql/OracleStoredProcedureTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 import org.hibernate.Session;
 import org.hibernate.dialect.OracleDialect;
 import org.hibernate.procedure.ProcedureCall;
-import org.hibernate.query.procedure.ProcedureParameter;
+import org.hibernate.procedure.ProcedureParameter;
 import org.hibernate.result.Output;
 import org.hibernate.result.ResultSetOutput;
 import org.hibernate.testing.orm.domain.userguide.Account;
